replace client api

+++ src/os/win32/os_file.c
@@ -18,7 +18,7 @@ s_sis_handle sis_open(const char *fn_, int mode_, int access_)
 	// printf("%s %x %x\n", fn_, mode_, access_);
 	return open(fn_, mode_, access_);
 }
-int64 sis_seek(s_sis_handle fp_, int64 offset_, int set_)
+long long sis_seek(s_sis_handle fp_, long long offset_, int set_)
 {
 	// return lseek(fp_, offset_, set_);
 	return _lseeki64(fp_, offset_, set_);
